Move disclose script tests to test_management_scripts
